requirements.txt
aioredis==1.0.0
asgiref==2.1.6
astroid==1.6.1
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
autopep8==1.3.4
channels==2.0.2
channels-redis==2.1.0
colorama==0.3.9
constantly==15.1.0
daphne==2.0.4
Django==2.0.2
django-cors-headers==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.5.6
msgpack-python==0.5.5
pycodestyle==2.3.1
PyJWT==1.5.3
pylint==1.8.2
pypiwin32==223
pytz==2018.3
pywin32==223
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
wrapt==1.10.11
zope.interface==4.4.3
